fix: type errors due to mapbox-gl v3.7.0
No longer augments the `Map` interface and the `Style` interface of the
`mapbox-gl` module. Instead, derives a new `Style` type from
`Map['style']`, which includes an optional `_getLayerSourceCache` method
that is provided in `mapbox-gl` v2.x.

The extracted interface of `Transform` is left incompatible with that of
`mapbox-gl` v3.7.0, so that it will be compatible with `mapbox-gl`
v3.8.0 or higher. Related type errors are suppressed with
`@ts-expect-error`

Decorates import statements with the `type` modifier, which import only
types definitions.

Turns `allowSyntheticDefaultImports` on in `tsconfig.json`, otherwise we
could not import the type declaration file of `mapbox-gl`.