fix: collectCollisionBoxesAndFeatures crashed with v3.9.0
Fixes the issue that `collectionCollisionBoxesAndFeatures` crashed with
`mapbox-gl` v3.9.0.

Since v3.9.0, `FeatureIndex.lookupSymbolFeatures` takes a `QrfQuery`
instead of `filterSpec` and `filterLayerIDs`. So adds a new version
indicator `isV3_9` which tells `mapbox-gl` is 3.9.0 or later, and uses
it to determine the parameters for `FeatureIndex.lookupSymbolFeatures`.
Utilizes the fact that the `Map` class has the `indoor` field since
v3.9.0.

A new file `src/private/qrf-query.ts` provides a utility function
`makeQrfQueryForStyleLayer` which creates a `QrfQuery` for a given
`StyleLayer`.

Derives the `StyleLayer` type based on the actual type extracted from
`mapbox-gl`, because the details of `StyleLayer` matter.