fix: type error with mapbox-gl v3.11.0
In `mapbox-gl` v3.11.0, it replaced the type of
`Style._availableImages`: `string[]` → `ImageId[]`. So rewrites the
parameter type of `FeatureIndex.lookupSymbolFeatures`. Replaces all the
overloaded declarations to suppress type errors. As declarations for
`mapbox-gl` v3.10.0 and earlier are also replaced, this treatment is not
precisely correct. However, since these declarations are internal, it
neither change behavior nor bother users of this library.