feat(gha): publish dev package to GitHub Packages
Introduces a GitHub Actions workflow which publishes a developer package
whenever commits are pushed to the `main` branch.

The `publish-dev-package` workflow starts when commits are pushed to the
`main` branch and calls the `publish-package` workflow that does the job
to publish a developer package.

A developer package has the version number followed by a dash (`-`) plus
the short commit hash of the commit used to build the package.